Updated packages and corresponding code
