+++ src/index.tsx
@@ -1,5 +1,6 @@
 import { render } from 'solid-js/web';
+
 import './index.css';
 import App from './App';
 
-render(() => <App />, document.getElementById("root"));
+render(() => <App/>, document.getElementById("root"));
