+++ src/pages/About.tsx
@@ -1,0 +1,11 @@
+import { Component } from 'solid-js';
+
+const About: Component = () => {
+  return (
+    <main class="text-center">
+      About page
+    </main>
+  );
+};
+
+export default About;
