+++ src/pages/Fallback.tsx
@@ -1,0 +1,11 @@
+import { Component } from 'solid-js';
+
+const Fallback: Component = () => {
+  return (
+    <main class="text-center">
+      Fallback Page
+    </main>
+  );
+};
+
+export default Fallback;
