+++ src/pages/Home.tsx
@@ -1,0 +1,11 @@
+import { Component } from 'solid-js';
+
+const Home: Component = () => {
+  return (
+    <main class="text-center">
+      Home Page
+    </main>
+  );
+};
+
+export default Home;
