--- src/views/About.tsx
@@ -1,61 +1,0 @@
-import { Component } from 'solid-js';
-import Header from '../components/Header'
-
-const About: Component = () => {
-  return (
-    <>
-      <Header title="About" />
-      <main>
-        <div class="max-w-7xl mx-auto py-6 sm:px-6 lg:px-8">
-          <div class="px-4 py-6 sm:px-0">
-            <div class="mt-10 mx-auto max-w-screen-xl px-4 sm:mt-12 sm:px-6 md:mt-16 lg:mt-20 lg:px-8 xl:mt-28">
-              <div class="sm:text-center lg:text-left">
-                <h2 class="text-4xl tracking-tight leading-10 font-extrabold text-gray-900 sm:text-5xl sm:leading-none md:text-6xl">
-                  This demo is using
-                  <br class="xl:hidden" />{' '}
-                  <a
-                    target="_blank"
-                    class="text-indigo-600 underline hover:text-indigo-500"
-                    href="https://tailwindcss.com"
-                  >
-                    Tailwind CSS
-                  </a>
-                </h2>
-                <p class="mt-3 text-base text-gray-500 sm:mt-5 sm:text-lg sm:max-w-xl sm:mx-auto md:mt-5 md:text-xl lg:mx-0">
-                  All the code present here is part of a free sample from{' '}
-                  <a
-                    href="https://tailwindui.com"
-                    class="text-indigo-600 underline hover:text-indigo-500"
-                  >
-                    Tailwind UI
-                  </a>{' '}
-                  but, you don't need it.
-                </p>
-                <div class="mt-5 sm:mt-8 sm:flex sm:justify-center lg:justify-start">
-                  <div class="rounded-md shadow">
-                    <a
-                      href="https://github.com/wobsoriano/vite-solid-tailwind-starter"
-                      class="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base leading-6 font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-500 focus:outline-none focus:ring transition duration-150 ease-in-out md:py-4 md:text-lg md:px-10"
-                    >
-                      Get started
-                    </a>
-                  </div>
-                  <div class="mt-3 sm:mt-0 sm:ml-3">
-                    <a
-                      href="#"
-                      class="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base leading-6 font-medium rounded-md text-indigo-700 bg-indigo-100 hover:text-indigo-600 hover:bg-indigo-50 focus:outline-none focus:ring focus:border-indigo-300 transition duration-150 ease-in-out md:py-4 md:text-lg md:px-10"
-                    >
-                      Live demo
-                    </a>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </main>
-    </>
-  );
-};
-
-export default About;
