--- src/views/NotFound.tsx
@@ -1,7 +1,0 @@
-import { Component } from 'solid-js';
-
-const NotFound: Component = () => {
-  return <main>Not found</main>
-};
-
-export default NotFound;
